Organise styled components like this
