+++ src/Pages/Blog.js
@@ -1,9 +1,9 @@
 import {
     React,
 } from "react";
-import styled from 'styled-components';
-import Navbar from '../Components/Navbar';
 import '../styles/global.css';
+import Navbar from '../Components/Navbar';
+import { Gen_Btn } from "../Components/Buttons.js";
 
 import { db } from '../index.js';
 import { 
@@ -19,29 +19,6 @@ import {
 } from "firebase/firestore";
 
 
-const theme = {
-    blue: {
-        default: "#0065fc",
-        hover: "#448efc",
-    },
-    red: {
-        default: "#fc1303",
-        hover: "#fa5246"
-    },
-};
-
-const Button = styled.button`
-    background-color: ${(props) => theme[props.theme].default};
-    padding: 5px 10px;
-    &:hover {
-        background-color: ${(props) => theme[props.theme].hover};
-    };
-`
-
-Button.defaultProps = {
-    theme: "blue",
-}
-
 async function getAllPosts() {
     const postsQuery = query(
         collection(db, 'posts'),
@@ -56,7 +33,7 @@ async function getAllPosts() {
 
 async function createPost() {
     let postDate = new Date().toLocaleString();
-    
+
     try {
         const docRef = await addDoc(collection(db, "comments"), {
             type: "comment",
@@ -76,10 +53,10 @@ function Blog(){
             <h1>Siiiiick DB Test</h1>
             <p>View in console, check posts on Firebase</p>
             <div>
-                <Button onClick={getAllPosts}>List Posts</Button>
+                <Gen_Btn onClick={getAllPosts}>List Posts</Gen_Btn>
             </div>
             <div>
-                <Button theme="red" onClick={createPost}>Create Post</Button>
+                <Gen_Btn theme="red" onClick={createPost}>Create Post</Gen_Btn>
             </div>
         </>
     )
